Update edit layout (#1501)
* Remove usage of size variables

From now on, every interface should render 100% wide. The edit form will
set the parent containers width according to the settings.

* Remove group interface

It's not officially supported yet, we don't have a GUI in settings to
set it up, and it doesn't work properly. We'll remove the group int
for the time being to avoid confusion.

* Set max len to 100

Lines in vue html are way easier to read when they're inline.
This change should allow enough wiggle room to clean up a lot of files

* Add optional string to en-us

Instead of having it hardcoded

* [WIP] Rework form / field to support new sizes

* Tweak prettier settings for an easier to read codebase

Vue html renders so much more legible this way, I think it improves
The readability of the codebase drastically, which should improve
ease of contribution.

* Fall back to default eslint rules in favor of prettierrc

* Release prettier

Auto layout the codebase based on the updated prettierrc config

* Fix conflict

* Fix right padding on wide info active

* Remove group component for the time being

It doesn't work well, was never officially supported and doesn't have
a gui to set it up...

* Implement form grid

half / full / fill options for interfaces

* Add responsiveness to form fields

* Clean up the SCSS

Split it up into separate files to increase readability.
I've also created a couple new variables that ensure we use the same
styles across the platform.

* Use v-icon on remaining icon instances

* Fix icons in wysiwyg

* Add event handlers for field

* Style batch toggle