Add dedicated JSON interface (#1533)
* Install jsonlint

* Move CodeMirror CSS into design folder

So it's being re-used by all the components that rely on codemirror

* Remove codemirror css from code interface

In favor of relying on the global codemirror styles

* Remove wrong import

I removed the code css file but forgot to delete the input

* Fix typo in the meta.json file

Both translations were called en-US. One was supposed to be fr-FR

* Add new JSON interface

It's using CodeMirror as well. It's loosely based on the code interface
except for the fact that it's way less flexible (by design). It's made
to handle JSON very well, and only that.

* Don't error on empty try catch statements